UserConfに保存の実装
